version 3
Trafico.py is a script that creates a JSON file from data of bluetooth traffic sensors in Vigo, Spain. You can download the code and change the number of iterations.
In an iteration, this script refresh the JSON file every five minutes and take the date and the total of all measure times at that moment.